+++ RELEASE-NOTES.md
@@ -1,5 +1,11 @@
 # qnd-papi-template release notes
 
+## 2026-02-12 - v0.0.4
+
+### Added/Refactoring/Deprecation
+
+- (Feat) Add API to return overview info of a symbol.
+
 ## 2026-02-11 - v0.0.3
 
 ### Added/Refactoring/Deprecation